+++ pkg/api/v1/handler.go
@@ -26,10 +26,7 @@ func TargetsHandler(targets []target.Client) http.HandlerFunc {
 // ClusterResourcesPolicyListHandler REST API
 func ClusterResourcesPolicyListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchClusterPolicies(Filter{
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-		})
+		list, err := finder.FetchClusterPolicies(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -37,11 +34,7 @@ func ClusterResourcesPolicyListHandler(finder PolicyReportFinder) http.HandlerFu
 // ClusterResourcesRuleListHandler REST API
 func ClusterResourcesRuleListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchClusterRules(Filter{
-			Sources:    req.URL.Query()["sources"],
-			Policies:   req.URL.Query()["policies"],
-			Categories: req.URL.Query()["categories"],
-		})
+		list, err := finder.FetchClusterRules(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -49,10 +42,7 @@ func ClusterResourcesRuleListHandler(finder PolicyReportFinder) http.HandlerFunc
 // NamespacedResourcesPolicyListHandler REST API
 func NamespacedResourcesPolicyListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchNamespacedPolicies(Filter{
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-		})
+		list, err := finder.FetchNamespacedPolicies(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -60,11 +50,7 @@ func NamespacedResourcesPolicyListHandler(finder PolicyReportFinder) http.Handle
 // NamespacedResourcesPolicyListHandler REST API
 func NamespacedResourcesRuleListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchNamespacedRules(Filter{
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Policies:   req.URL.Query()["policies"],
-		})
+		list, err := finder.FetchNamespacedRules(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -72,9 +58,7 @@ func NamespacedResourcesRuleListHandler(finder PolicyReportFinder) http.HandlerF
 // CategoryListHandler REST API
 func CategoryListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchCategories(Filter{
-			Sources: req.URL.Query()["sources"],
-		})
+		list, err := finder.FetchCategories(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -82,12 +66,7 @@ func CategoryListHandler(finder PolicyReportFinder) http.HandlerFunc {
 // ClusterResourcesKindListHandler REST API
 func ClusterResourcesKindListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchClusterKinds(Filter{
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Policies:   req.URL.Query()["policies"],
-			Rules:      req.URL.Query()["rules"],
-		})
+		list, err := finder.FetchClusterKinds(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -95,13 +74,7 @@ func ClusterResourcesKindListHandler(finder PolicyReportFinder) http.HandlerFunc
 // NamespacedResourcesKindListHandler REST API
 func NamespacedResourcesKindListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchNamespacedKinds(Filter{
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Policies:   req.URL.Query()["policies"],
-			Rules:      req.URL.Query()["rules"],
-			Namespaces: req.URL.Query()["namespaces"],
-		})
+		list, err := finder.FetchNamespacedKinds(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -109,13 +82,7 @@ func NamespacedResourcesKindListHandler(finder PolicyReportFinder) http.HandlerF
 // ClusterResourcesListHandler REST API
 func ClusterResourcesListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchClusterResources(Filter{
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Policies:   req.URL.Query()["policies"],
-			Rules:      req.URL.Query()["rules"],
-			Kinds:      req.URL.Query()["kinds"],
-		})
+		list, err := finder.FetchClusterResources(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -123,14 +90,7 @@ func ClusterResourcesListHandler(finder PolicyReportFinder) http.HandlerFunc {
 // NamespacedResourcesListHandler REST API
 func NamespacedResourcesListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchNamespacedResources(Filter{
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Policies:   req.URL.Query()["policies"],
-			Rules:      req.URL.Query()["rules"],
-			Namespaces: req.URL.Query()["namespaces"],
-			Kinds:      req.URL.Query()["kinds"],
-		})
+		list, err := finder.FetchNamespacedResources(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -151,18 +111,26 @@ func NamespacedSourceListHandler(finder PolicyReportFinder) http.HandlerFunc {
 	}
 }
 
+// NamespacedReportLabelListHandler REST API
+func NamespacedReportLabelListHandler(finder PolicyReportFinder) http.HandlerFunc {
+	return func(w http.ResponseWriter, req *http.Request) {
+		list, err := finder.FetchNamespacedReportLabels(buildFilter(req))
+		helper.SendJSONResponse(w, list, err)
+	}
+}
+
+// ClusterReportLabelListHandler REST API
+func ClusterReportLabelListHandler(finder PolicyReportFinder) http.HandlerFunc {
+	return func(w http.ResponseWriter, req *http.Request) {
+		list, err := finder.FetchClusterReportLabels(buildFilter(req))
+		helper.SendJSONResponse(w, list, err)
+	}
+}
+
 // ClusterResourcesStatusCountHandler REST API
 func ClusterResourcesStatusCountHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchStatusCounts(Filter{
-			Kinds:      req.URL.Query()["kinds"],
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Severities: req.URL.Query()["severities"],
-			Policies:   req.URL.Query()["policies"],
-			Rules:      req.URL.Query()["rules"],
-			Status:     req.URL.Query()["status"],
-		})
+		list, err := finder.FetchStatusCounts(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -170,16 +138,7 @@ func ClusterResourcesStatusCountHandler(finder PolicyReportFinder) http.HandlerF
 // NamespacedResourcesStatusCountsHandler REST API
 func NamespacedResourcesStatusCountsHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		list, err := finder.FetchNamespacedStatusCounts(Filter{
-			Namespaces: req.URL.Query()["namespaces"],
-			Kinds:      req.URL.Query()["kinds"],
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Severities: req.URL.Query()["severities"],
-			Policies:   req.URL.Query()["policies"],
-			Rules:      req.URL.Query()["rules"],
-			Status:     req.URL.Query()["status"],
-		})
+		list, err := finder.FetchNamespacedStatusCounts(buildFilter(req))
 		helper.SendJSONResponse(w, list, err)
 	}
 }
@@ -198,19 +157,7 @@ func RuleStatusCountHandler(finder PolicyReportFinder) http.HandlerFunc {
 // NamespacedResourcesResultHandler REST API
 func NamespacedResourcesResultHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		filter := Filter{
-			Namespaces: req.URL.Query()["namespaces"],
-			Kinds:      req.URL.Query()["kinds"],
-			Resources:  req.URL.Query()["resources"],
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Severities: req.URL.Query()["severities"],
-			Policies:   req.URL.Query()["policies"],
-			Rules:      req.URL.Query()["rules"],
-			Status:     req.URL.Query()["status"],
-			Search:     req.URL.Query().Get("search"),
-		}
-
+		filter := buildFilter(req)
 		count, err := finder.CountNamespacedResults(filter)
 		list, err := finder.FetchNamespacedResults(filter, buildPaginatiomn(req))
 		helper.SendJSONResponse(w, ResultList{Items: list, Count: count}, err)
@@ -220,16 +167,7 @@ func NamespacedResourcesResultHandler(finder PolicyReportFinder) http.HandlerFun
 // ClusterResourcesResultHandler REST API
 func ClusterResourcesResultHandler(finder PolicyReportFinder) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		filter := Filter{
-			Kinds:      req.URL.Query()["kinds"],
-			Resources:  req.URL.Query()["resources"],
-			Sources:    req.URL.Query()["sources"],
-			Categories: req.URL.Query()["categories"],
-			Severities: req.URL.Query()["severities"],
-			Policies:   req.URL.Query()["policies"],
-			Rules:      req.URL.Query()["rules"],
-			Status:     req.URL.Query()["status"],
-		}
+		filter := buildFilter(req)
 		count, err := finder.CountClusterResults(filter)
 		list, err := finder.FetchClusterResults(filter, buildPaginatiomn(req))
 		helper.SendJSONResponse(w, ResultList{Items: list, Count: count}, err)
@@ -272,5 +210,32 @@ func buildPaginatiomn(req *http.Request) Pagination {
 		Offset:    offset,
 		SortBy:    sortBy,
 		Direction: direction,
+	}
+}
+
+func buildFilter(req *http.Request) Filter {
+	labels := map[string]string{}
+
+	for _, label := range req.URL.Query()["labels"] {
+		parts := strings.Split(label, ":")
+		if len(parts) != 2 {
+			continue
+		}
+
+		labels[strings.TrimSpace(parts[0])] = strings.TrimSpace(parts[1])
+	}
+
+	return Filter{
+		Namespaces:  req.URL.Query()["namespaces"],
+		Kinds:       req.URL.Query()["kinds"],
+		Resources:   req.URL.Query()["resources"],
+		Sources:     req.URL.Query()["sources"],
+		Categories:  req.URL.Query()["categories"],
+		Severities:  req.URL.Query()["severities"],
+		Policies:    req.URL.Query()["policies"],
+		Rules:       req.URL.Query()["rules"],
+		Status:      req.URL.Query()["status"],
+		ReportLabel: labels,
+		Search:      req.URL.Query().Get("search"),
 	}
 }