+++ pkg/listener/metrics/result_custom_test.go
@@ -13,10 +13,11 @@ import (
 )
 
 func Test_CustomResultMetricGeneration(t *testing.T) {
-	gauge := metrics.RegisterCustomResultGauge("policy_report_custom_result", []string{"namespace", "policy", "status", "source"})
+	gauge := metrics.RegisterCustomResultGauge("policy_report_custom_result", []string{"namespace", "policy", "status", "source", "app"})
 
 	report1 := report.PolicyReport{
 		ID:                "1",
+		Labels:            map[string]string{"app": "policy-reporter"},
 		Name:              "polr-test",
 		Namespace:         "test",
 		Summary:           report.Summary{Pass: 2, Fail: 1},
@@ -26,6 +27,7 @@ func Test_CustomResultMetricGeneration(t *testing.T) {
 
 	report2 := report.PolicyReport{
 		ID:                "1",
+		Labels:            map[string]string{"app": "policy-reporter"},
 		Name:              "polr-test",
 		Namespace:         "test",
 		Summary:           report.Summary{Pass: 0, Fail: 1},
@@ -36,7 +38,7 @@ func Test_CustomResultMetricGeneration(t *testing.T) {
 	filter := metrics.NewResultFilter(validate.RuleSets{}, validate.RuleSets{}, validate.RuleSets{Exclude: []string{"disallow-policy"}}, validate.RuleSets{}, validate.RuleSets{})
 
 	t.Run("Added Metric", func(t *testing.T) {
-		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source"}))
+		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source", "label:app"}, []string{"namespace", "policy", "status", "source", "app"}))
 		handler(report.LifecycleEvent{Type: report.Added, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
 
 		metricFam, err := prometheus.DefaultGatherer.Gather()
@@ -61,7 +63,7 @@ func Test_CustomResultMetricGeneration(t *testing.T) {
 	t.Run("Modified Metric", func(t *testing.T) {
 		gauge.Reset()
 
-		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source"}))
+		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source", "label:app"}, []string{"namespace", "policy", "status", "source", "app"}))
 		handler(report.LifecycleEvent{Type: report.Added, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
 		handler(report.LifecycleEvent{Type: report.Updated, NewPolicyReport: report2, OldPolicyReport: report1})
 
@@ -76,10 +78,10 @@ func Test_CustomResultMetricGeneration(t *testing.T) {
 		}
 
 		metrics := results.GetMetric()
-		if err = testCustomResultMetricLabels(metrics[0], result2, 0); err != nil {
-			t.Error(err)
+		if len(metrics) != 1 {
+			t.Fatalf("Expected only one metric is left, got %d\n", len(metrics))
 		}
-		if err = testCustomResultMetricLabels(metrics[1], result1, 1); err != nil {
+		if err = testCustomResultMetricLabels(metrics[0], result1, 1); err != nil {
 			t.Error(err)
 		}
 	})
@@ -87,7 +89,7 @@ func Test_CustomResultMetricGeneration(t *testing.T) {
 	t.Run("Deleted Metric", func(t *testing.T) {
 		gauge.Reset()
 
-		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source"}))
+		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source", "label:app"}, []string{"namespace", "policy", "status", "source", "app"}))
 		handler(report.LifecycleEvent{Type: report.Added, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
 		handler(report.LifecycleEvent{Type: report.Updated, NewPolicyReport: report2, OldPolicyReport: report1})
 		handler(report.LifecycleEvent{Type: report.Deleted, NewPolicyReport: report2, OldPolicyReport: report.PolicyReport{}})
@@ -98,15 +100,37 @@ func Test_CustomResultMetricGeneration(t *testing.T) {
 		}
 
 		results := findMetric(metricFam, "policy_report_custom_result")
+		if results != nil {
+			t.Fatalf(" expected metric policy_report_custom_result no longer exists")
+		}
+	})
+
+	t.Run("Decrease Metric", func(t *testing.T) {
+		gauge.Reset()
+
+		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source", "label:app"}, []string{"namespace", "policy", "status", "source", "app"}))
+		handler(report.LifecycleEvent{Type: report.Added, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
+		handler(report.LifecycleEvent{Type: report.Added, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
+		handler(report.LifecycleEvent{Type: report.Deleted, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
+
+		metricFam, err := prometheus.DefaultGatherer.Gather()
+		if err != nil {
+			t.Errorf("unexpected Error: %s", err)
+		}
+
+		results := findMetric(metricFam, "policy_report_custom_result")
 		if results == nil {
 			t.Fatalf("Metric not found: policy_report_custom_result")
 		}
 
 		metrics := results.GetMetric()
-		if err = testCustomResultMetricLabels(metrics[0], result2, 0); err != nil {
+		if len(metrics) != 2 {
+			t.Fatalf("Expected only one metric is left, got %d\n", len(metrics))
+		}
+		if err = testCustomResultMetricLabels(metrics[0], result2, 1); err != nil {
 			t.Error(err)
 		}
-		if err = testCustomResultMetricLabels(metrics[1], result1, 0); err != nil {
+		if err = testCustomResultMetricLabels(metrics[1], result1, 1); err != nil {
 			t.Error(err)
 		}
 	})
@@ -114,6 +138,15 @@ func Test_CustomResultMetricGeneration(t *testing.T) {
 
 func testCustomResultMetricLabels(metric *ioprometheusclient.Metric, result report.Result, expVal float64) error {
 	var index int
+
+	if name := *metric.Label[index].Name; name != "app" {
+		return fmt.Errorf("unexpected App Label: %s", name)
+	}
+	if value := *metric.Label[index].Value; value != "policy-reporter" {
+		return fmt.Errorf("unexpected Namespace Label Value: %s", value)
+	}
+
+	index++
 
 	if name := *metric.Label[index].Name; name != "namespace" {
 		return fmt.Errorf("unexpected Name Label: %s", name)