+++ pkg/target/client.go
@@ -1,6 +1,9 @@
 package target
 
 import (
+	"strings"
+
+	"github.com/kyverno/go-wildcard"
 	"github.com/kyverno/policy-reporter/pkg/helper"
 	"github.com/kyverno/policy-reporter/pkg/report"
 	"github.com/kyverno/policy-reporter/pkg/validate"
@@ -14,15 +17,15 @@ type Client interface {
 	SkipExistingOnStartup() bool
 	// Name is a unique identifier for each Target
 	Name() string
-	// Validate is a result should send
-	Validate(result report.Result) bool
+	// Validate if a result should send
+	Validate(rep report.PolicyReport, result report.Result) bool
 	// MinimumPriority for a triggered Result to send to this target
 	MinimumPriority() string
 	// Sources of the Results which should send to this target, empty means all sources
 	Sources() []string
 }
 
-func NewClientFilter(namespace, priority, policy validate.RuleSets, minimumPriority string, sources []string) *report.ResultFilter {
+func NewResultFilter(namespace, priority, policy validate.RuleSets, minimumPriority string, sources []string) *report.ResultFilter {
 	f := report.NewResultFilter()
 	f.Sources = sources
 	f.MinimumPriority = minimumPriority
@@ -60,16 +63,65 @@ func NewClientFilter(namespace, priority, policy validate.RuleSets, minimumPrior
 	return f
 }
 
+func NewReportFilter(labels validate.RuleSets) *report.ReportFilter {
+	f := report.NewReportFilter()
+	if labels.Count() > 0 {
+		f.AddValidation(func(r report.PolicyReport) bool {
+			if len(labels.Include) > 0 {
+				for _, label := range labels.Include {
+					parts := strings.Split(label, ":")
+					if len(parts) == 1 {
+						parts = append(parts, "*")
+					}
+
+					labelName := strings.TrimSpace(parts[0])
+					labelValue := strings.TrimSpace(parts[1])
+
+					for key, value := range r.Labels {
+						if labelName == key && wildcard.Match(labelValue, value) {
+							return true
+						}
+					}
+				}
+
+				return false
+			} else if len(labels.Exclude) > 0 {
+				for _, label := range labels.Exclude {
+					parts := strings.Split(label, ":")
+					if len(parts) == 1 {
+						parts = append(parts, "*")
+					}
+
+					labelName := strings.TrimSpace(parts[0])
+					labelValue := strings.TrimSpace(parts[1])
+
+					for key, value := range r.Labels {
+						if labelName == key && wildcard.Match(labelValue, value) {
+							return false
+						}
+					}
+				}
+			}
+
+			return true
+		})
+	}
+
+	return f
+}
+
 type BaseClient struct {
 	name                  string
 	skipExistingOnStartup bool
-	filter                *report.ResultFilter
+	resultFilter          *report.ResultFilter
+	reportFilter          *report.ReportFilter
 }
 
 type ClientOptions struct {
 	Name                  string
 	SkipExistingOnStartup bool
-	Filter                *report.ResultFilter
+	ResultFilter          *report.ResultFilter
+	ReportFilter          *report.ReportFilter
 }
 
 func (c *BaseClient) Name() string {
@@ -77,27 +129,31 @@ func (c *BaseClient) Name() string {
 }
 
 func (c *BaseClient) MinimumPriority() string {
-	if c.filter == nil {
+	if c.resultFilter == nil {
 		return report.DefaultPriority.String()
 	}
 
-	return c.filter.MinimumPriority
+	return c.resultFilter.MinimumPriority
 }
 
 func (c *BaseClient) Sources() []string {
-	if c.filter == nil {
+	if c.resultFilter == nil {
 		return make([]string, 0)
 	}
 
-	return c.filter.Sources
+	return c.resultFilter.Sources
 }
 
-func (c *BaseClient) Validate(result report.Result) bool {
-	if c.filter == nil {
-		return true
+func (c *BaseClient) Validate(rep report.PolicyReport, result report.Result) bool {
+	if c.reportFilter != nil && !c.reportFilter.Validate(rep) {
+		return false
 	}
 
-	return c.filter.Validate(result)
+	if c.resultFilter != nil && !c.resultFilter.Validate(result) {
+		return false
+	}
+
+	return true
 }
 
 func (c *BaseClient) SkipExistingOnStartup() bool {
@@ -105,5 +161,5 @@ func (c *BaseClient) SkipExistingOnStartup() bool {
 }
 
 func NewBaseClient(options ClientOptions) BaseClient {
-	return BaseClient{options.Name, options.SkipExistingOnStartup, options.Filter}
+	return BaseClient{options.Name, options.SkipExistingOnStartup, options.ResultFilter, options.ReportFilter}
 }