+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+# 2.14.0
+* Policy Reporter
+    * Persist also PolicyReport labels
+    * API
+        * New API to get available labels for PolicyReports: `/v1/namespaced-resources/report-labels`
+        * New API to get available labels for ClusterPolicyReports: `/v1/cluster-resources/report-labels`
+    * Metrics
+        * special syntax to add report labels to metric labels: `label:report-label-name`, special characters like `-`, `/`, `.`, `:` will be transformed to `_` in metrics
+    * New Target Filter `reportLabel` to, filter results based on labels of the related (Cluster)PolicyReport
+* Monitoring
+    * New values to disable dedicated Grafana Dashboards:
+        - `grafana.dashboards.enable.overview`, default `true`
+        - `grafana.dashboards.enable.policyReportDetails`, default `true`
+        - `grafana.dashboards.enable.clusterPolicyReportDetails`, default `true`
+    * New values to configure the Grafana Dashboard datasource label, pluginName, pluginId
+        - `grafana.datasource.label`, default `Prometheus`
+        - `grafana.datasource.pluginName`, default `Prometheus`
+        - `grafana.datasource.pluginId`, default `prometheus`
+    * New value `grafana.dashboards.labelFilter` to add custom report labels as dashboard filter, default `[]`. Label has to be a valid prometheus label, e.g. `created-by` => `created_by`.
+    * New values `grafana.dashboards.multicluster.enabled` and `grafana.dashboards.multicluster.label` to add an optional `cluster` label.
+
+
 # 2.13.5
 * Add configuration `target.s3.pathStyle` for the `S3` output